# PERF: significant speedups in tz-aware operations