--- a/pandas/core/arrays/datetimes.py
+++ b/pandas/core/arrays/datetimes.py
@@ -335,7 +335,9 @@ class DatetimeArrayMixin(dtl.DatetimeLikeArrayMixin,
             cls._validate_frequency(result, freq, ambiguous=ambiguous)
 
         elif freq_infer:
-            result.freq = to_offset(result.inferred_freq)
+            # Set _freq directly to bypass duplicative _validate_frequency
+            # check.
+            result._freq = to_offset(result.inferred_freq)
 
         return result