@mroeschke 2018-12-30
Overall LGTM. Could you also add `tzlocal` in the relevant benchmarks `that parameterize over timezones?