@jbrockmendel 2018-12-30
There are I think 7 places in tslibs where we do this iteration.  Can they all be optimized?