@jbrockmendel 2018-12-30
Another perf improvement waiting in these files is calling searchsorted via `cnp.PyArray_SearchSorted` instead of `arr.searchsorted`, since the latter is a python-space call.  Doing this is at least 3 kinds of a hassle so I never implemented it myself.  Feel free to ping me if you're interested in giving it a go.