@jbrockmendel 2018-12-30
> a bit -1 on actually using the numpy c-api for searchsorted as it adds complexity; I suspect it won't actually speed things up for us (and is out of scope for this PR anyhow).

Definitely agree on the scope, and like I said it is a PITA.  But it's also one of the few python-space calls left in these files that we do at each stage of a loop.