@qwhelan 2019-01-03
@jreback Good to merge, can't reproduce the spurious regression.